Naming: cast no longer necessary here.
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@7869 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ hpx/runtime/naming/address.hpp
@@ -104,7 +104,7 @@ namespace hpx { namespace naming
     {
         boost::io::ios_flags_saver ifs(os);
         os << "(" << addr.locality_ << ":"
-           << components::get_component_type_name(static_cast<int>(addr.type_))
+           << components::get_component_type_name(addr.type_)
            << ":" << std::showbase << std::hex << addr.address_ << ")";
         return os;
     }
